+++ Backend/Interfaces/IMemoryAccessor.cs
@@ -2,8 +2,8 @@ namespace Backend.Interfaces
 {
     public interface IMemoryAccessor : IDisposable
     {
-        int ReadInt32(int address);
-        short ReadInt16(int address);
-        void ReadArray(int address, byte[] buffer, int index, int count);
+        int ReadInt32(long address);
+        short ReadInt16(long address);
+        void ReadArray(long address, byte[] buffer, int index, int count);
     }
 }
