+++ Backend/Interfaces/IMemoryReaderService.cs
@@ -4,9 +4,9 @@ namespace Backend.Interfaces
     {
         bool IsConnected { get; }
         bool TryConnect();
-        byte[]? ReadBytes(int address, int length);
-        int ReadInt32(int address);
-        short ReadInt16(int address);
-        string ReadString(int address, int length, System.Text.Encoding? encoding = null);
+        byte[]? ReadBytes(long address, int length);
+        int ReadInt32(long address);
+        short ReadInt16(long address);
+        string ReadString(long address, int length, System.Text.Encoding? encoding = null);
     }
 }
